+++ src/netzob/Common/Models/Types/TypeConverter.py
@@ -77,28 +77,28 @@ class TypeConverter(object):
         To convert a raw data to its decimal representation and then to its ASCII representation
 
         >>> data = '\x23'
-        >>> decData = TypeConverter.convert(data, Raw, Decimal)
+        >>> decData = TypeConverter.convert(data, Raw, Integer)
         >>> print decData
         35
-        >>> print TypeConverter.convert(decData, Decimal, ASCII)
+        >>> print TypeConverter.convert(decData, Integer, ASCII)
         #
 
         You can also play with the unitSize to convert multiple ascii in a single high value decimal
 
-        >>> TypeConverter.convert("5", ASCII, Decimal)
+        >>> TypeConverter.convert("5", ASCII, Integer)
         53
-        >>> print TypeConverter.convert("zoby", ASCII, Decimal)
+        >>> print TypeConverter.convert("zoby", ASCII, Integer)
         2036494202
-        >>> print TypeConverter.convert("zoby", ASCII, Decimal, dst_unitSize=AbstractType.UNITSIZE_32)
+        >>> print TypeConverter.convert("zoby", ASCII, Integer, dst_unitSize=AbstractType.UNITSIZE_32)
         2054120057
 
         It also works for 'semantic' data like IPv4s
 
-        >>> TypeConverter.convert("192.168.0.10", IPv4, Decimal, dst_sign=AbstractType.SIGN_UNSIGNED)
+        >>> TypeConverter.convert("192.168.0.10", IPv4, Integer, dst_sign=AbstractType.SIGN_UNSIGNED)
         167815360
         >>> TypeConverter.convert("127.0.0.1", IPv4, BitArray)
         bitarray('01111111000000000000000000000001')
-        >>> TypeConverter.convert(167815360, Decimal, IPv4, src_unitSize=AbstractType.UNITSIZE_32, src_sign=AbstractType.SIGN_UNSIGNED)
+        >>> TypeConverter.convert(167815360, Integer, IPv4, src_unitSize=AbstractType.UNITSIZE_32, src_sign=AbstractType.SIGN_UNSIGNED)
         IPAddress('10.0.168.192')
 
         :param sourceType: the data source type
