+++ src/netzob/Common/Models/Vocabulary/Functions/EncodingFunctions/DomainEncodingFunction.py
@@ -59,8 +59,8 @@ class DomainEncodingFunction(EncodingFunction):
     which domains is based on the aggregation of two other variables.
 
     >>> from netzob.all import *
-    >>> f = Field(name="f0", domain=Agg(["There are ", Decimal(10), " solutions."]))
-    >>> m = RawMessage("There are " + TypeConverter.convert(10, Decimal, Raw) + " solutions.")
+    >>> f = Field(name="f0", domain=Agg(["There are ", Integer(10), " solutions."]))
+    >>> m = RawMessage("There are " + TypeConverter.convert(10, Integer, Raw) + " solutions.")
     >>> s = Symbol(fields=[f], messages=[m], name="Symbol")
     >>> s.addEncodingFunction(TypeEncodingFunction(ASCII))
     >>> print s
